refactor(theme/boards): create Boards TabularItem.vue & MoreOptionsDropdown.vue
